+++ odbc-api/src/preallocated.rs
@@ -1,10 +1,6 @@
 use widestring::{U16Str, U16String};
 
-use crate::{
-    execute::{execute_columns, execute_with_parameters},
-    handles::StatementImpl,
-    CursorImpl, Error, ParameterCollection,
-};
+use crate::{CursorImpl, Error, ParameterCollection, execute::{execute_columns, execute_tables, execute_with_parameters}, handles::StatementImpl};
 
 /// A preallocated SQL statement handle intended for sequential execution of different queries. See
 /// [`crate::Connection::preallocate`].
@@ -109,6 +105,41 @@ impl<'o> Preallocated<'o> {
     /// accessible through safe abstractions.
     pub fn into_statement(self) -> StatementImpl<'o> {
         self.statement
+    }
+
+    /// List tables, schemas, views and catalogs of a datasource.
+    /// 
+    /// # Parameters
+    /// 
+    /// * `catalog_name`: Filter result by catalog name. Accept search patterns. Use `%` to match
+    ///   any number of characters. Use `_` to match exactly on character. Use `\` to escape
+    ///   characeters.
+    /// * `schema_name`: Filter result by schema. Accepts patterns in the same way as
+    ///   `catalog_name`.
+    /// * `table_name`: Filter result by table. Accepts patterns in the same way as `catalog_name`.
+    /// * `table_type`: Filters results by table type. E.g: 'TABLE', 'VIEW'. This argument accepts a
+    ///   comma separeted list of table types. Omit it to not filter the result by table type at
+    ///   all.
+    pub fn tables(
+        &mut self,
+        catalog_name: Option<&str>,
+        schema_name: Option<&str>,
+        table_name: Option<&str>,
+        table_type: Option<&str>,
+    ) -> Result<CursorImpl<&mut  StatementImpl<'o>>, Error> {
+
+        let catalog_name = catalog_name.map(|s| U16String::from_str(s));
+        let schema_name = schema_name.map(|s| U16String::from_str(s));
+        let table_name = table_name.map(|s| U16String::from_str(s));
+        let table_type = table_type.map(|s| U16String::from_str(s));
+
+        execute_tables(
+            &mut self.statement,
+            catalog_name.as_deref(),
+            schema_name.as_deref(),
+            table_name.as_deref(),
+            table_type.as_deref(),
+        )
     }
 
     /// Query all columns that match the provided catalog name, schema pattern, table pattern, and